+++ packages/client-utils/src/index.ts
@@ -6,5 +6,6 @@ export { mapKeyringTransaction } from './mappers/keyring-transaction-mapper.js';
 export { mapLocalTransaction } from './mappers/local-transaction-mapper.js';
 export { mapRampsOrder } from './mappers/ramps-order-mapper.js';
 export type { RampsOrderLike } from './mappers/ramps-order-mapper.js';
+export type { GetKnownTokenDecimals } from './mappers/helpers/transactions.js';
 
 export type * from './types.js';